pyproject.toml
[build-system]
requires = ["setuptools==75.1.0", "wheel==0.44.0"]
build-backend = "setuptools.build_meta"

[project]
name = "python-orchestra"
version = "0.1.0"
description = "Agentic code review loop orchestrator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai==1.63.2",
  "requests==2.32.3",
  "PyYAML==6.0.2",
  "pytest==8.3.4",
]

[tool.pytest.ini_options]
addopts = "-q"
